refactor(codex): extract procgroup package
