+++ internal/procgroup/process_group_unix.go
@@ -1,6 +1,6 @@
 //go:build unix
 
-package codex
+package procgroup
 
 import (
 	"errors"
@@ -8,17 +8,17 @@ import (
 	"syscall"
 )
 
-func configureCommandProcessGroup(cmd *exec.Cmd) {
+func Configure(cmd *exec.Cmd) {
 	if cmd.SysProcAttr == nil {
 		cmd.SysProcAttr = &syscall.SysProcAttr{}
 	}
 	cmd.SysProcAttr.Setpgid = true
 	cmd.Cancel = func() error {
-		return terminateCommandProcessTree(cmd, commandProcessGroupID(cmd))
+		return TerminateTree(cmd, GroupID(cmd))
 	}
 }
 
-func commandProcessGroupID(cmd *exec.Cmd) int {
+func GroupID(cmd *exec.Cmd) int {
 	if cmd == nil || cmd.Process == nil {
 		return 0
 	}
@@ -29,7 +29,7 @@ func commandProcessGroupID(cmd *exec.Cmd) int {
 	return pgid
 }
 
-func terminateCommandProcessTree(cmd *exec.Cmd, processGroupID int) error {
+func TerminateTree(cmd *exec.Cmd, processGroupID int) error {
 	if processGroupID > 0 {
 		err := syscall.Kill(-processGroupID, syscall.SIGKILL)
 		if err == nil || errors.Is(err, syscall.ESRCH) {
@@ -43,7 +43,7 @@ func terminateCommandProcessTree(cmd *exec.Cmd, processGroupID int) error {
 	return cmd.Process.Kill()
 }
 
-func cleanupCommandProcessGroup(processGroupID int) error {
+func Cleanup(processGroupID int) error {
 	if processGroupID <= 0 {
 		return nil
 	}
